refactor: split i18n.rs into i18n/ module (types.rs + mod.rs)
